+++ Tests/MapboxMapsTests/Style/CustomSourcesTests.swift
@@ -1,4 +1,3 @@
-// This file is generated.
 import XCTest
 @_spi(Experimental) @testable import MapboxMaps
 
@@ -8,6 +7,7 @@ final class CustomSourcesSourceTests: XCTestCase {
         let testCustomRasterSourceOptions = CustomRasterSourceOptions(fetchTileFunction: { _ in }, cancelTileFunction: { _ in })
 
         var source = CustomRasterSource(id: "test-source", options: testCustomRasterSourceOptions)
+        source.tileCacheBudget = TileCacheBudgetSize.testSourceValue(TileCacheBudgetSize.megabytes(7))
 
         var data: Data?
         do {
@@ -25,6 +25,7 @@ final class CustomSourcesSourceTests: XCTestCase {
             let decodedSource = try JSONDecoder().decode(CustomRasterSource.self, from: validData)
             XCTAssert(decodedSource.type == SourceType.customRaster)
             XCTAssert(decodedSource.id == "test-source")
+            XCTAssert(decodedSource.tileCacheBudget == TileCacheBudgetSize.testSourceValue(TileCacheBudgetSize.megabytes(7)))
             XCTAssertNil(decodedSource.options)
         } catch {
             XCTFail("Failed to decode CustomRasterSource.")
@@ -35,6 +36,7 @@ final class CustomSourcesSourceTests: XCTestCase {
         let testCustomGeometrySourceOptions = CustomGeometrySourceOptions(fetchTileFunction: { _ in }, cancelTileFunction: { _ in }, tileOptions: TileOptions())
 
         var source = CustomGeometrySource(id: "test-source", options: testCustomGeometrySourceOptions)
+        source.tileCacheBudget = TileCacheBudgetSize.testSourceValue()
 
         var data: Data?
         do {
@@ -52,11 +54,10 @@ final class CustomSourcesSourceTests: XCTestCase {
             let decodedSource = try JSONDecoder().decode(CustomGeometrySource.self, from: validData)
             XCTAssert(decodedSource.type == SourceType.customGeometry)
             XCTAssert(decodedSource.id == "test-source")
+            XCTAssert(decodedSource.tileCacheBudget == TileCacheBudgetSize.testSourceValue())
             XCTAssertNil(decodedSource.options)
         } catch {
             XCTFail("Failed to decode CustomRasterSource.")
         }
     }
 }
-
-// End of generated file